Migrated Telestaff Roster Data to a singleton model
